requirements.txt
# =============================================================
# LLM-Judge Semantic Uncertainty - Dependencies
# =============================================================

# --- Core ---
openai>=1.30.0
datasets>=2.19.0              # HuggingFace TriviaQA 로드
pyyaml>=6.0

# --- NLP / NER ---
spacy>=3.7.0                  # Answer Type 분류 (NER)

# --- Data & Analysis ---
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0                 # 통계 검정

# --- Visualization ---
matplotlib>=3.9.0
seaborn>=0.13.0

# --- Async / Rate Limiting ---
aiohttp>=3.9.0
tenacity>=8.3.0               # retry with backoff
tqdm>=4.66.0                  # 프로그레스 바

# --- Environment ---
python-dotenv>=1.0.0

# --- Dev / Test ---
pytest>=8.2.0
ipykernel>=6.29.0             # Jupyter 커널
